Adds integration tests
The integration tests will build the charm from scratch, and deploy
it in a currently bootstrapped Juju Controller in a new test model.

The integration tests will also deploy the postgresql-k8s and
nginx-ingress-integrator charms and relates them to the finos-waltz-k8s
charm. The tests are asserting that Waltz can be reached through its IP
and through its default service-hostname.