actions: Removing the explicit controller name
The demo environment was updated from juju 3.0 to 3.1. For this, a new
controller had to be bootstrapped, and the waltz-public-v3 model was
migrated to it.

Since we only have one controller, we don't need to explicitly specify it.
This will make it easier in the future, when upgrading the controller once again.