actions: Adds manua db refresh period and update database dump actions
The update-refresh-period GitHub action will allow someone to update the
Waltz demo instance's database refresh period to a custom value (in minutes).
Setting it to 0 will disable the database refresh.

The set-database-dump GitHub action will allow someone to update the
SQL database dump to be used by the Waltz instance.